Safari-611.1.10/LayoutTests/http/tests/dom/noopener-window-cannot-target-opener-expected.txt

CONSOLE MESSAGE: PASS: New window should not have an opener
CONSOLE MESSAGE: PASS: New window should have URL 'about:blank'
Make sure that windows opened with 'noopener' via window.open cannot target their opener.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS w is null
